pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["anand-kamble <anandmk837@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["standard"], version = "^0.115.6"}
uvicorn = {extras = ["standard"], version = "^0.32.1"}
commitizen = "^4.0.0"
supabase = "^2.10.0"
python-dotenv = "^1.0.1"
langgraph = "^0.2.56"
langsmith = "^0.1.147"
langchain-openai = "^0.2.11"
langchain-anthropic = "^0.3.0"
psycopg-pool = "^3.2.4"
langgraph-checkpoint-postgres = "^2.0.7"
psycopg = {extras = ["binary", "pool"], version = "^3.2.3"}
langchain-community = "^0.3.9"


[tool.poetry.scripts]
start = "backend:start"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
